fixed bugs

+++ src/components/Packages/Packages.jsx
@@ -26,28 +26,80 @@ export default function Card(props) {
   const [item, setItem] = useState({
     name: "",
     price: 0,
-    pax: 0,
+    pax: 1,
     dates: "",
     image: "",
   });
 
-  const handleChange = async (e, name, price, image) => {
-    console.log("Entrando a HC");
+  const handleChange = (e, name, price, image) => {
     setItem({
       ...item,
       [e.target.name]: e.target.value,
-      name: name,
-      image: image,
-      price: price,
+      name,
+      price,
+      image
     });
   };
 
-  const handleClick = (event) => {
+  const handleClickPreCart = (name, price, image) => {
+    setItem({
+      ...item,
+      name: name,
+      image: image,
+      price: price,
+    })
+  }
+
+  const handleClick = () => {
+    if (document.getElementById(`${item.name} passengers`).value <= 0) {
+      return Swal.fire({
+        title: "You must add at least one passenger",
+        text: item.name,
+        imageUrl: item.image,
+        imageWidth: 400,
+        imageHeight: 200,
+        imageAlt: "Custom image",
+      });
+    }
+    if (document.getElementById(`${item.name} dates`).value === 'select') {
+      return Swal.fire({
+        title: "You must select a date to continue",
+        text: item.name,
+        imageUrl: item.image,
+        imageWidth: 400,
+        imageHeight: 200,
+        imageAlt: "Custom image",
+      });
+    }
+
     let arrayItemsStore = JSON.parse(localStorage.getItem("items"));
+    // console.log(document.getElementById('dates').value)
     if (arrayItemsStore === null) arrayItemsStore = [];
+    if (arrayItemsStore.find(e => e.name === item.name && e.dates === item.dates)) {
+      document.getElementById(`${item.name} dates`).value = 'select'
+      return Swal.fire({
+        title: "You already have this item in your cart",
+        text: item.name,
+        imageUrl: item.image,
+        imageWidth: 400,
+        imageHeight: 200,
+        imageAlt: "Custom image",
+      });
+      
+    }
+    if (!arrayItemsStore.find(e => e.name === item.name && e.dates === item.dates)) {
     arrayItemsStore.push(item);
     localStorage.setItem("items", JSON.stringify(arrayItemsStore));
-    Swal.fire({
+    // Alert
+    document.getElementById(`${item.name} dates`).value = 'select'
+    setItem({
+      name: "",
+      price: 0,
+      pax: 1,
+      dates: "",
+      image: "",
+    });
+    return Swal.fire({
       title: "Added to cart successfully!",
       text: item.name,
       imageUrl: item.image,
@@ -55,14 +107,8 @@ export default function Card(props) {
       imageHeight: 200,
       imageAlt: "Custom image",
     });
-    setItem({
-      name: "",
-      price: 0,
-      pax: 0,
-      dates: "",
-      image: "",
-    });
-  };
+  }
+};
 
   //   Fin Precart
 
@@ -153,12 +199,12 @@ export default function Card(props) {
                           {/* Boton Modal */}
                           <button
                             type="button"
+                            onClick={() => handleClickPreCart(e.name, e.price, e.image)}
                             className="btn btn-outline-secondary btn-lg"
                             data-bs-toggle="modal"
                             data-bs-target={`#${e.name.split(" ").join("")}`}
                           >
-                            <i className="bi bi-cart-check"></i> Add to my
-                            trips!
+                            <i className="bi bi-cart-check"></i> Add to Cart!
                           </button>
                           {/* Fin Boton Modal */}
                         </div>
@@ -226,7 +272,7 @@ export default function Card(props) {
                                   <div className="row ">
                                     <div className="col-md-8">
                                       <p className="text-end">
-                                        Please, choose the number of passengers:
+                                        Please, select the number of passengers:
                                       </p>
                                     </div>
                                     <div className="col-md-4 text-start">
@@ -234,7 +280,7 @@ export default function Card(props) {
                                         className={styles.cantpackages}
                                         name="pax"
                                         min="1"
-                                        id="pax"
+                                        id={`${item.name} passengers`}
                                         type="number"
                                         value={item.pax}
                                         defaultValue="1"
@@ -244,7 +290,6 @@ export default function Card(props) {
                                             e.name,
                                             e.price,
                                             e.image,
-                                            e.dates
                                           )
                                         }
                                       />
@@ -253,7 +298,7 @@ export default function Card(props) {
                                   <div className="row ">
                                     <div className="col-md-8">
                                       <p className="text-end">
-                                        Please, choose date:
+                                        Please, select a date:
                                       </p>
                                     </div>
                                     <div className="col-md-4 text-start">
@@ -264,19 +309,16 @@ export default function Card(props) {
                                             e.name,
                                             e.price,
                                             e.image,
-                                            e.dates
                                           )
                                         }
                                         name="dates"
-                                        id="dates"
+                                        id={`${e.name} dates`}
                                       >
-                                        <option disabled selected>
-                                          Choose
+                                        <option value="select" disabled selected>
+                                          Select
                                         </option>
-                                        {e.dates.split(",").map((e) => {
-                                          return (
-                                            <option value={e}>{e} </option>
-                                          );
+                                        {e.dates?.split(",").map((e) => {
+                                          return <option value={e}>{e} </option>;
                                         })}
                                       </select>
                                     </div>
@@ -292,7 +334,7 @@ export default function Card(props) {
                                       >
                                         TOTAL:{" "}
                                         <i className="bi bi-currency-dollar"></i>
-                                        ARS {e.price * item.pax}
+                                        ARS {e.price * item.pax < 0 ? 0 : e.price * item.pax}
                                       </div>
                                     </div>
                                   </div>
@@ -309,8 +351,8 @@ export default function Card(props) {
                                   Cancel
                                 </button>
                                 <button
-                                  onClick={(event) => {
-                                    handleClick(event);
+                                  onClick={() => {
+                                    handleClick();
                                   }}
                                   type="button"
                                   className="btn btn-primary"
